+++ src/AppserverIo/Appserver/Core/AbstractManager.php
@@ -26,6 +26,7 @@ use AppserverIo\Storage\GenericStackable;
 use AppserverIo\Psr\Application\ManagerInterface;
 use AppserverIo\Psr\Application\ApplicationInterface;
 use AppserverIo\Psr\Naming\InitialContext as NamingDirectory;
+use AppserverIo\Appserver\ServletEngine\RequestHandler;
 
 /**
  * Abstract manager implementation.
@@ -159,7 +160,18 @@ abstract class AbstractManager extends GenericStackable implements ManagerInterf
      */
     public function lookupProxy($lookupName, $sessionId = null)
     {
-        return $this->getInitialContext()->lookup($lookupName, $sessionId);
+
+        // load the initial context instance
+        $initialContext = $this->getInitialContext();
+
+        // query whether a request context is available
+        if ($servletRequest = RequestHandler::getRequestContext()) {
+            // inject the servlet request to handle SFSBs correctly
+            $initialContext->injectServletRequest($servletRequest);
+        }
+
+        // lookup the proxy by the name and session ID if available
+        return $initialContext->lookup($lookupName, $sessionId);
     }
 
     /**